pyproject.toml
[tool.poetry]
name = "geojsonapi"
version = "0.1.0"
description = "GeoJSON API"
authors = ["José Vitor Santos Silva <jvss23031999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.3"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
django-email-users = "^1.1.0"
djangorestframework = "^3.14.0"
pytest = "^7.2.0"
pytest-env = "^0.8.1"
django-extensions = "^3.2.1"
pycountry = "^22.3.5"
drf-spectacular = "^0.24.2"
pytest-lazy-fixture = "^0.6.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
flake8 = "^6.0.0"
flake8-pyproject = "^1.2.1"
pytest = "^7.2.0"
pytest-sugar = "^0.9.6"
pytest-mock = "^3.10.0"
pytest-xdist = "^3.0.2"
pytest-cov = "^4.0.0"
coverage = "^6.5.0"
black = "^22.10.0"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py", "tests.py"]
env = [
    "DJANGO_SETTINGS_MODULE=geojsonapi.settings",
]
filterwarnings = [
     "ignore:The --rsyncdir command line argument and rsyncdirs config variable are deprecated.:DeprecationWarning",
]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.black]
experimental-string-processing = true
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |\.tox
  |\.env
)/
'''

[tool.flake8]
max-line-length = 120
